requirements.txt
setuptools

httpx~=0.28.1

pathlib~=1.0.1
ujson~=5.10.0
numpy~=1.26.4

uvicorn~=0.34.0
fastapi~=0.115.12
pydantic~=2.11.2
websockets~=15.0.1
jinja2
python-multipart
ujson
soundfile
aiofiles
pydub~=0.25.1

tqdm
psutil

# For ASR
#-f https://k2-fsa.github.io/sherpa/onnx/cuda.html
#sherpa-onnx==1.11.3+cuda
onnxruntime-gpu # ==1.17.1
librosa

# For punctuation
transformers~=4.51.0

# For diarisation
scikit-learn==1.5.2  # Понижен, чтоыбы избегать варнингов
python_speech_features
umap-learn
hdbscan


# for VAD
requests~=2.32.3

# For monitoring
pynvml~=12.0.0
nvidia-ml-py~=12.570.86
